pseudo modularizing bessel
